third_party/blink/web_tests/external/wpt/custom-elements/registries/adoption.window-expected.txt

This is a testharness.js-based test.
[FAIL] Adoption with global registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with explicit global registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with scoped registry
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with global registry into a scoped registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with explicit global registry into a scoped registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with scoped registry into a scoped registry
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with global registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with explicit global registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with scoped registry
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with global registry into a scoped registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with explicit global registry into a scoped registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with scoped registry into a scoped registry
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with global registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with explicit global registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with scoped registry (null registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with global registry into a scoped registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with explicit global registry into a scoped registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption with scoped registry into a scoped registry (null registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with global registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with explicit global registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with scoped registry (null registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with global registry into a scoped registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with explicit global registry into a scoped registry (null registry target)
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Adoption including shadow root with scoped registry into a scoped registry (null registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with global registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with explicit global registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with global registry into a scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with explicit global registry into a scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption with scoped registry into a scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with global registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with explicit global registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with global registry into a scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with explicit global registry into a scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Adoption including shadow root with scoped registry into a scoped registry (scoped registry target)
  Failed to construct 'CustomElementRegistry': Illegal constructor
Harness: the test ran to completion.
